% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{code_script}
\alias{code_script}
\title{A dataset to be used to practice with the porfolio analysis functions.}
\format{
\subsection{\code{code_script}}{

A data frame with 5 rows and 5 columns:
\describe{
\item{fund}{Name of the fund (character)}
\item{vintage_year}{Year the fund began investing (integer)}
\item{commitment}{Total capital committed in USD (numeric)}
\item{called}{Capital called to date in USD (numeric)}
\item{recallable}{Recallable distributions in USD (numeric)}
}
}
}
\usage{
code_script
}
\description{
A dataset to be used to practice with the porfolio analysis functions.
}
\keyword{datasets}
