% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfunded_liability_schedule.R
\name{unfunded_liability_schedule}
\alias{unfunded_liability_schedule}
\title{Calculate Unfunded PE Liabilities and Create an Appropriate Schedule}
\usage{
unfunded_liability_schedule(data)
}
\arguments{
\item{data}{A data frame containing information about the fund. Column names default to fund, commitment, called, recallable.}
}
\value{
A tibble showing current unfunded liabilities and a 5-year deployment schedule.
}
\description{
Calculate Unfunded PE Liabilities and Create an Appropriate Schedule
}
\examples{
df <- data.frame(
  fund = c("Fund A", "Fund B"),
  commitment = c(10000000, 15000000),
  called = c(4000000, 3000000),
  recallable = c(1000000, 500000)
)
unfunded_liability_schedule(df)
}
